third_party/blink/web_tests/animations/prefixed/animation-events-unprefixed-04.html

<!DOCTYPE html>
<title>Tests that custom events with unprefixed animations names are correctly dispatched.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
var animationStartReceived = false;
var animationIterationReceived = false;

document.addEventListener('animationstart', () => {
  animationStartReceived = true;
});

document.addEventListener('webkitAnimationStart', () => {
  assert_unreached("webkitAnimationStart event listener should not have been called");
});

document.addEventListener('animationiteration', () => {
  animationIterationReceived = true;
});

document.addEventListener('webkitAnimationIteration', () => {
  assert_unreached("webkitAnimationIteration event listener should not have been called");
});

document.addEventListener('webkitAnimationEnd', function(e) {
  assert_unreached("webkitAnimationEnd event listener should not have been called");
});

async_test(t => {
  document.addEventListener('animationend', () => {
    t.step(() => {
      assert_true(animationStartReceived && animationIterationReceived);
      t.done();
    });
  });

  var custom = document.createEvent('CustomEvent');
  custom.initCustomEvent('animationstart', true, true, null);
  document.dispatchEvent(custom);
  custom = document.createEvent('CustomEvent');
  custom.initCustomEvent('animationiteration', true, true, null);
  document.dispatchEvent(custom);
  custom = document.createEvent('CustomEvent');
  custom.initCustomEvent('animationend', true, true, null);
  document.dispatchEvent(custom);
}, "Tests that custom events with unprefixed animations names are correctly dispatched");
</script>
